+++ src/utils/storage.js
@@ -1,0 +1,22 @@
+const storage = window.localStorage
+
+export const setItem = (key, value) => {
+    try {
+        storage.setItem(key, JSON.stringify(value))
+    } catch (e) {
+        console.log(e)
+    }
+}
+
+export const getItem = (key, defaultValue) => {
+    try {
+        const storedValue = storage.getItem(key)
+        if (storedValue) {
+            return JSON.parse(storedValue)
+        }
+        return defaultValue
+    } catch (e) {
+        console.log(e)
+        return defaultValue
+    }
+}
